refactor(userpage): move APIs into g0v0 private API
