+++ app/router/private/__init__.py
@@ -1,6 +1,6 @@
 from app.config import settings
 
-from . import admin, audio_proxy, avatar, beatmapset, cover, oauth, relationship, score, team, username  # noqa: F401
+from . import admin, audio_proxy, avatar, beatmapset, cover, oauth, relationship, score, team, user  # noqa: F401
 from .router import router as private_router
 
 if settings.enable_totp_verification:
